requirements.txt
torch>=2.0
gymnasium>=0.29
pygame>=2.5
numpy>=1.24
matplotlib>=3.7
streamlit>=1.30
wandb>=0.16
pyyaml>=6.0
stable-baselines3>=2.2
plotly>=5.18
imageio>=2.33
